# openapi/admin-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chaos Admin API",
    "version": "1.0.0",
    "description": "Passwordless administration API for merchant accounts and headless commerce stores."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin"
    }
  ],
...
  "paths": {
...
    "/admin/v1/stores/{store_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/StoreId"
        }
      ],
      "get": {
        "operationId": "getStore",
        "tags": [
          "Stores"
        ],
        "summary": "Get Store configuration",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Store detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updateStore",
        "tags": [
          "Stores"
        ],
        "summary": "Update Store configuration",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStore"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResourceMutated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
    "/admin/v1/stores/{store_id}/activate": {
      "parameters": [
        {
          "$ref": "#/components/parameters/StoreId"
        },
        {
          "$ref": "#/components/parameters/IdempotencyKey"
        }
      ],
      "post": {
        "operationId": "activateStore",
        "tags": [
          "Stores"
        ],
        "summary": "Activate a ready Store",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResourceMutated"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
    "/admin/v1/stores/{store_id}/archive": {
      "parameters": [
        {
          "$ref": "#/components/parameters/StoreId"
        },
        {
          "$ref": "#/components/parameters/IdempotencyKey"
        }
      ],
      "post": {
        "operationId": "archiveStore",
        "tags": [
          "Stores"
        ],
        "summary": "Archive a Store",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResourceMutated"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/admin/v1/stores/{store_id}/sales-channels": {
      "parameters": [
        {
          "$ref": "#/components/parameters/StoreId"
        }
      ],
      "get": {
        "operationId": "listSalesChannels",
        "tags": [
          "Stores"
        ],
        "summary": "List Sales Channels",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Sales Channel page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesChannelCollectionEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      },
      "post": {
        "operationId": "createSalesChannel",
        "tags": [
          "Stores"
        ],
        "summary": "Create an active Sales Channel",
        "security": [
          {
            "bearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteSalesChannel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sales Channel created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentifierResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque revocable session token"
      }
    },
    "parameters": {
      "StoreId": {
        "name": "store_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque cursor returned by the previous page.",
        "schema": {
          "type": "string"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed JSON, query, content type, or payload size",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or credential invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated caller is not allowed to access the resource",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist in the authorized scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency or uniqueness conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ValidationFailed": {
        "description": "Domain or request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "ResourceMutated": {
        "description": "Resource mutation completed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/IdentifierResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "UpdateStore": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "name",
          "default_region",
          "default_currency"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{0,30}[a-z0-9]$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "default_region": {
            "$ref": "#/components/schemas/RegionCode"
          },
          "default_currency": {
            "$ref": "#/components/schemas/CurrencyCode"
          }
        }
      },
      "SalesChannelKind": {
        "type": "string",
        "enum": [
          "web",
          "mobile",
          "point_of_sale",
          "marketplace",
          "custom"
        ]
      },
...
      "WriteSalesChannel": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "name",
          "kind"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{0,30}[a-z0-9]$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "kind": {
            "$ref": "#/components/schemas/SalesChannelKind"
          }
        }
      },
...
      "RegionCode": {
        "type": "string",
        "pattern": "^[A-Z]{2}$",
        "default": "US"
      },
      "CurrencyCode": {
        "type": "string",
        "pattern": "^[A-Z]{3}$",
        "default": "USD"
      },
...
      "IdentifierResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Identifier"
          }
        }
      },
...
      "StoreEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Store"
          }
        }
      },
...
      "SalesChannelCollectionEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SalesChannel"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
      "ResponseMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "page"
        ],
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    }
  }
}